Implement the splits.io API
This brings the splits.io API to livesplit-core via an optional
`networking` module that is disabled by default. Additional helpers that
directly upload and download runs are provided as well as the raw API
bindings.
