+++ Cargo.toml
@@ -74,13 +74,13 @@ vek = { version = "0.9.8", default-features = false, optional = true }
 web-sys = { version = "0.3.28", default-features = false, features = ["Performance", "Window"], optional = true }
 
 # Networking
-splits-io-api = { git = "https://github.com/LiveSplit/splits-io-api", optional = true }
+splits-io-api = { version = "0.1.0", optional = true }
 
 [dev-dependencies]
 memmem = "0.1.1"
 criterion = "0.3.0"
 crc = "1.8.1"
-tokio = "0.2.0-alpha.4"
+tokio = { version = "0.2", features = ["macros"] }
 
 [features]
 default = ["image-shrinking", "std"]
