+++ src/networking/mod.rs
@@ -1,1 +1,6 @@
+//! The networking module provides functionality to communicate with various
+//! speedrunning related websites, such as splits.io to upload and download runs
+//! and speedrun.com to query and submit to the leaderboards of most games. The
+//! module is optional and is not compiled in by default.
+
 pub mod splits_io;
